+++ git/capture_output/capture_output.go
