+++ git/capture_output/capture_output_test.go
