+++ subcommands/mktmp/mktmp/mktmp.go
@@ -13,7 +13,6 @@ func CheckoutTmp(tree string) (string, error) {
 
 	repo, err := git.OpenRepository(".")
 	if err != nil {
-		fmt.Println(err.Class)
 		return "", err
 	}
