+++ subcommands/mktmp/tbd_mktmp.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/tbd-ci/tbd/tbd_mktmp/mktmp"
+	"github.com/tbd-ci/tbd/subcommands/mktmp/mktmp"
 )
 
 func init() {
