+++ subcommands/note/tbd_note.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"regexp"
 
-	"github.com/tbd-ci/tbd/tbd_note/note"
+	"github.com/tbd-ci/tbd/subcommands/note/note"
 )
 
 func init() {
