+++ subcommands/note/tbd_note.man
