+++ subcommands/run/tbd_run.go
@@ -8,9 +8,9 @@ import (
 	"flag"
 	"fmt"
 	git "github.com/libgit2/git2go"
-	"github.com/tbd-ci/tbd/capture_output"
-	"github.com/tbd-ci/tbd/empty_ref"
-	"github.com/tbd-ci/tbd/nested_write"
+	"github.com/tbd-ci/tbd/git/capture_output"
+	"github.com/tbd-ci/tbd/git/empty_ref"
+	"github.com/tbd-ci/tbd/git/nested_write"
 	"os"
 	"os/exec"
 	"syscall"
