+++ subcommands/status/tbd_status.go
